+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 # Version 1.1.1+6 RC
 
  - Added mutex to back end so no device shows up twice.
- - Fixed dialog theme not showing in light mode.  
+ - Fixed dialog theme not showing in light mode.
+ - Reverted Vive Base station report back to UNKNOWN, still needs some research before this works.
 
 # Version 1.1.0+5 30-09-2020