+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 
-# Version 1.1.1+6 RC
+# Version 1.1.1+6 02-10-2020
 
  - Added mutex to back end so no device shows up twice.
  - Fixed dialog theme not showing in light mode.